Fix bug in script for refreshing debug log

+++ scripts/clear_log.sh
@@ -1,2 +1,4 @@
-rm debug.log
+if test -f "debug.log"; then
+  rm debug.log
+fi
 touch debug.log
